Move `define_units()` into `helpers.py` file
